Refactor repositories and services from WebApp submodule to separate submodules
