+++ Services/src/main/java/ev/projects/services/ICaseService.java
@@ -1,13 +1,13 @@
-package ev.projects.webapp.repositories;
+package ev.projects.services;
 
 import ev.projects.models.Case;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ICaseRepository {
+public interface ICaseService {
 
     List<Case> getAll();
-    Optional<Case> getById(long ID);
+    Optional<Case> getAllById(Long ID);
 
 }
