+++ Services/src/main/java/ev/projects/services/IDocumentService.java
@@ -1,4 +1,4 @@
-package ev.projects.webapp.services;
+package ev.projects.services;
 
 import ev.projects.models.Document;
