+++ WebApp/src/main/java/ev/projects/webapp/WebAppApplication.java
@@ -2,8 +2,10 @@ package ev.projects.webapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan({"ev.projects.services", "ev.projects.webapp"})
 public class WebAppApplication {
 
     public static void main(String[] args) {
