+++ WebApp/src/main/java/ev/projects/webapp/restControllers/CaseController.java
@@ -1,7 +1,7 @@
 package ev.projects.webapp.restControllers;
 
 import ev.projects.models.Case;
-import ev.projects.webapp.services.ICaseService;
+import ev.projects.services.ICaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
